Move the nullness-operator tests to their own class
Rename the sixteen passing tests after the property each establishes, now
that the class name carries the issue context. The nine ignored ones keep
the issue prefix, since their names double as the checklist for removing
the annotations.

Add an ignored test for a dereference of a type variable with a nullable
upper bound, which reproduces the defect without a wildcard.
